Added full flow test variants for CSB and GSB to prevent code duplication
