+++ frontend/e2e-tests/test-data/eml-files.ts
@@ -67,25 +67,25 @@ export const eml110a_AB: Eml110a = {
   path: "../backend/src/eml/tests/eml110a_test_AB.eml.xml",
   electionName: "Waterschap Rivier en Polder 2023",
   electionDate: "woensdag 15 maart 2023",
-  hashInput1: "7c7c",
-  hashInput2: "98b5",
+  hashInput1: "2989",
+  hashInput2: "d0f4",
   fullHash: [
-    "826e",
-    "01a7",
-    "0264",
-    "3575",
-    "2784",
-    "fa1b",
-    "4448",
-    "7c7c",
-    "a90e",
-    "494e",
-    "a96f",
-    "3090",
-    "8c1c",
-    "98b5",
-    "02c1",
-    "aeb8",
+    "f958",
+    "1366",
+    "c63f",
+    "b36e",
+    "2989",
+    "bf4e",
+    "4bb2",
+    "e6a8",
+    "ba2f",
+    "e9b6",
+    "9d40",
+    "7ac6",
+    "f546",
+    "863c",
+    "6a6c",
+    "d0f4",
   ],
 };
 
@@ -278,7 +278,7 @@ export const eml230b_more_than_45_candidates: Eml230b = {
 
 export const eml230b_AB_more_than_45_candidates: Eml230b = {
   filename: "eml230b_test_AB_more_than_45_candidates.eml.xml",
-  path: "../backend/src/eml/tests/eml230b_test_AB_more_than_45_candidates.eml.xml",
+  path: "e2e-tests/test-data/eml-files/eml230b_test_AB_more_than_45_candidates.eml.xml",
   electionDate: "woensdag 15 maart 2023",
   hashInput1: "28bd",
   hashInput2: "1f07",
